// src/types.ts
export enum ContentCategory {
    HEADER = "HEADER",
    FOOTER = "FOOTER",
    HERO = "HERO",
    ABOUT = "ABOUT",
    SERVICES = "SERVICES",
    TEXT = "TEXT",
    CONTACT = "CONTACT"
}

export interface ProjectStruct {
    id: string
    title: string
    pages: Record<string, PageStruct>
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface PageStruct {
    id: string
    path: string
    description: string
    content: ContentStruct[]
    contentActive: ContentStruct | null
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface ContentStruct {
    id: string
    description: string
    Component: React.FC | string
    categories: ContentCategory[]
    props: any
}

/**
 * Colours
 */
export interface ColourScheme {
    /**
     * This is the main color that represents the brand
     * or the most dominant color in the website's palette. 
     * It's used for major call-to-action (CTA) buttons, 
     * links, and headlines to draw attention.
     */
    primary: string,

    /**
     * Supports the primary color and is used for secondary buttons, 
     * highlights, and information boxes. 
     * The secondary color should complement the primary color.
     */
    secondary: string,

    /**
     * Lighter shade of the primary color
     * These are additional colors used sparingly to accentuate
     * or highlight parts of the website.
     * They can be used for icons,
     * info graphics, or to highlight key information.
     */
    accent: string,

    /**
     * A fixed light grey
     * These are usually neutral colors that
     * don't compete with the primary and secondary colors.
     */
    background: string,


    /**
     * A fixed dark grey for text
     * Besides the obvious black or white,
     * you might have specific colors for headings,
     * subheadings, and body text that fit within
     * the overall color scheme.
     */
    text: string,

    /**
     * Lighter shade of the secondary color
     * These include shades of gray, beige,
     * or off-white used for text, background,
     * borders, and shadows.
     * They help create hierarchy and contrast in the design.
     */
    neutral: string,

    palette: string[]
}

export enum ColourPalette {
    ACCENT = '--color-accent',
    BACKGROUND = '--color-background',
    NEUTRAL = '--color-neutral',
    PRIMARY = '--color-primary',
    SECONDARY = '--color-secondary',
    TEXT = '--color-text'
}


/**
 * Font
 */
export interface FamilyFont {
    name: string
    family: string
    css: string
    description?: string
}

export interface FontScheme {
    heading: FamilyFont
    body: FamilyFont
}
// ...
export enum PageMessageType {
    SET_ID = "SET_ID",
    SET_PATH = "SET_PATH",
    SET_DESCRIPTION = "SET_DESCRIPTION",

    SET_CONTENT = "SET_CONTENT",
    PUT_CONTENT = "PUT_CONTENT",
    SET_CONTENT_ACTIVE = "SET_CONTENT_ACTIVE",

    SET_COLOUR_SCHEME = "SET_COLOUR_SCHEME",
    SET_FONT_SCHEME = "SET_FONT_SCHEME",

    NOTIFY_CONTENT_UPDATED = "NOTIFY_CONTENT_UPDATED",
}

export type PageEvent = {
    type: PageMessageType.SET_PATH
    data: string
} | {
    type: PageMessageType.SET_ID
    data: string
} | {
    type: PageMessageType.SET_DESCRIPTION
    data: string
} | {
    type: PageMessageType.SET_CONTENT | PageMessageType.PUT_CONTENT
    data: PageContent[]
} | {
    type: PageMessageType.SET_CONTENT_ACTIVE
    data: PageContent | null
} | {
    type: PageMessageType.SET_COLOUR_SCHEME
    data: ColourScheme
} | {
    type: PageMessageType.SET_FONT_SCHEME
    data: FontScheme
} | {
    type: PageMessageType.NOTIFY_CONTENT_UPDATED
}